fix(core): Support execution recovery when saving execution progress
https://linear.app/n8n/issue/PAY-1697
